KEA-83 move locale/language stuff to config
